initial vectorized / block aggregators

+++ processing/src/test/java/io/druid/query/timeboundary/TimeBoundaryQueryTest.java
@@ -80,7 +80,7 @@ public class TimeBoundaryQueryTest
     );
 
 
-    Assert.assertEquals(1, serdeQuery.getContextValue("priority"));
+    Assert.assertEquals(1L, (long)serdeQuery.getContextValue("priority"));
     Assert.assertEquals(true, serdeQuery.getContextValue("useCache"));
     Assert.assertEquals(true, serdeQuery.getContextValue("populateCache"));
     Assert.assertEquals(true, serdeQuery.getContextValue("finalize"));
